refactor(*): split library and cli
