+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/gladiaio/gladia-cli
+
+go 1.18
+
+require (
+	github.com/olekukonko/tablewriter v0.0.5
+	golang.org/x/text v0.14.0
+)
+
+require github.com/mattn/go-runewidth v0.0.9 // indirect
