--- src/go.mod
@@ -1,9 +1,0 @@
-module gladia
-
-go 1.20
-
-require (
-	github.com/mattn/go-runewidth v0.0.9 // indirect
-	github.com/olekukonko/tablewriter v0.0.5 // indirect
-	golang.org/x/text v0.14.0 // indirect
-)
